Entry point refactoring
