+++ src/index.ts
@@ -1,43 +1,12 @@
-import { app, protocol, BrowserWindow, Tray, Menu, MenuItemConstructorOptions, MenuItem } from 'electron';
-import { createProtocol } from 'vue-cli-plugin-electron-builder/lib';
+import { app, protocol, BrowserWindow } from 'electron';
 import installExtension, { VUEJS_DEVTOOLS } from 'electron-devtools-installer';
 import notify from './utils/notification';
+import createWindow from '@/utils/AppWindow';
 
-/**
- * Tray element
- */
-let tray: Tray;
-
-/**
- *  Creates the system menu that will be displayed by right-arrow click on the app icon.
- *
- *  @returns { Menu }
- */
-function createAppMenu(): Menu {
-  /**
-   * Menu element creating
-   */
-  const template: (MenuItemConstructorOptions | MenuItem)[] = [
-    {
-      label: 'About',
-      role: 'about',
-    },
-    {
-      label: 'Quit',
-      role: 'quit',
-    },
-  ];
-
-  return Menu.buildFromTemplate(template);
-}
 /**
  * Node environment
  */
 const isDevelopment = process.env.NODE_ENV !== 'production';
-/**
- * Protocol name
- */
-const protocolName = 'app';
 
 /**
  * Scheme must be registered before the app is ready
@@ -51,61 +20,6 @@ protocol.registerSchemesAsPrivileged([
     },
   },
 ]);
-
-/**
- * Creating window
- */
-async function createWindow(): Promise<void> {
-  const win = new BrowserWindow({
-    height: 352,
-    width: 260,
-    frame: false,
-    resizable: false,
-    show: false,
-    transparent: true,
-    vibrancy: 'dark',
-    visualEffectState: 'active',
-    webPreferences: {
-      nodeIntegration: true,
-      enableRemoteModule: true,
-    },
-  });
-
-  if (process.env.WEBPACK_DEV_SERVER_URL) {
-    await win.loadURL(process.env.WEBPACK_DEV_SERVER_URL);
-  } else {
-    createProtocol(protocolName);
-    await win.loadURL(`${protocolName}://./index.html`);
-  }
-  const iconName = process.platform === 'win32' ? 'front.png' : 'front-mac.png';
-  const iconPath = `src/assets/images/${iconName}`;
-
-  tray = new Tray(iconPath);
-  tray.setIgnoreDoubleClickEvents(true);
-  tray.on('click', (event, bounds) => {
-    const { x, y } = bounds;
-    const { height, width } = win.getBounds();
-
-    if (win.isVisible()) {
-      win.hide();
-    } else {
-      const yPosition = process.platform === 'darwin' ? y : y - height;
-
-      win.setBounds({
-        x: Math.round(x - width / 2),
-        y: yPosition,
-        height,
-        width,
-      });
-      win.show();
-    }
-  });
-  const menu = createAppMenu();
-
-  tray.on('right-click', () => {
-    tray.popUpContextMenu(menu);
-  });
-}
 
 /**
  * Quit when all windows are closed.
